backend/recomm/requirements.txt
absl-py==1.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2022.9.24
charset-normalizer==2.1.1
clang==5.0
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djongo==1.3.6
entrypoints==0.4
executing==1.1.0
flatbuffers==2.0.7
fonttools==4.37.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
joblib==1.2.0
JPype1==1.4.0
jupyter-core==4.11.1
jupyter_client==7.3.5
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.21.5
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pymongo==3.12.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.2.4
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.4.0
typing-extensions==3.7.4.3
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.8.1
